Move DiaryForm to its own folder and add basic styles
