+++ src/components/DiaryForm/styles.css
@@ -1,0 +1,22 @@
+form {
+  max-width: 350px;
+  display: grid;
+  grid-template-columns: 1;
+  text-align: center;
+  font-size: 1rem;
+}
+
+label {
+  margin: 1rem 0 4px 0.5rem;
+  font-weight: bold;
+}
+
+input,
+select,
+button {
+  height: 30px;
+}
+
+button {
+  margin-top: 40px;
+}
